Add Middleware superclass
