+++ src/middleware/authorization_handler.py
@@ -1,18 +1,16 @@
-from requests.adapters import HTTPAdapter
-
-from ._base import AuthProviderBase
+from ._base_auth import AuthProviderBase
 from .options.constants import AUTH_MIDDLEWARE_OPTIONS
+from ..core._middleware import Middleware
 
 
-class AuthorizationHandler(HTTPAdapter):
+class AuthorizationHandler(Middleware):
     def __init__(self, auth_provider: AuthProviderBase, auth_provider_options=None):
         super().__init__()
 
         self.auth_provider = auth_provider
         self.auth_provider_options = auth_provider_options
-        self.next = None
 
-    def send(self, request, stream=False, timeout=None, verify=True, cert=None, proxies=None):
+    def send(self, request, **kwargs):
         scopes = self._get_middleware_options(request)
 
         if scopes:
@@ -21,12 +19,9 @@ class AuthorizationHandler(HTTPAdapter):
         token = self.auth_provider.get_access_token()
         request.headers.update({'Authorization': 'Bearer {}'.format(token)})
 
-        if self.next is None:
-            return super().send(request, stream, timeout, verify, cert, proxies)
-
-        return self.next.send(request, stream, timeout, verify, cert, proxies)
+        return super().send(request, **kwargs)
 
     def _get_middleware_options(self, request):
         options = request.middleware_control.get(AUTH_MIDDLEWARE_OPTIONS) or self.auth_provider_options
-        return options.scopes
+        return ''
